% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-doRedisContainer.R
\name{configWorkerContainerEnv,doRedisContainer-method}
\alias{configWorkerContainerEnv,doRedisContainer-method}
\title{Configure the doRedis worker container}
\usage{
\S4method{configWorkerContainerEnv}{doRedisContainer}(container, cluster, workerNumber, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{doRedisContainer} object
}
\description{
Configure the doRedis worker container
}
