% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-doRedisContainer.R
\name{deregisterParallelBackend,doRedisContainer-method}
\alias{deregisterParallelBackend,doRedisContainer-method}
\title{Deregister the foreach doRedis backend}
\usage{
\S4method{deregisterParallelBackend}{doRedisContainer}(container, cluster, verbose = FALSE, ...)
}
\arguments{
\item{container}{The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}

\item{...}{Not used}
}
\value{
No return value
}
\description{
Deregister the foreach doRedis backend. This will register the sequential backend.
}
