% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{doRedisContainer-class}
\alias{doRedisContainer-class}
\alias{.doRedisContainer}
\title{The doRedis container object}
\description{
The doRedis container object, it is used by the DockerParallel package to create
the worker container
}

