% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-RedisContainer.R
\name{doRedisServerContainer}
\alias{doRedisServerContainer}
\title{Get the Bioconductor Redis server container}
\usage{
doRedisServerContainer(environment = list(), tag = "latest")
}
\arguments{
\item{environment}{List, the environment variables in the container}

\item{tag}{Character, the image tag}
}
\value{
a \code{foreachRedisContainer} object
}
\description{
Get the Bioconductor Redis server container.
}
\examples{
RedisServerContainer()
}
