% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-RedisContainer.R
\name{doRedisWorkerContainer}
\alias{doRedisWorkerContainer}
\title{Get the doRedis worker container}
\usage{
doRedisWorkerContainer(
  image = c("r-base", "bioconductor"),
  RPackages = NULL,
  sysPackages = NULL,
  environment = list(),
  maxWorkerNum = 4L,
  tag = "latest"
)
}
\arguments{
\item{image}{Character, the worker image used by the container}

\item{RPackages}{Character, a vector of R packages that will be installed
by \code{AnVIL::install} before connecting with the server}

\item{sysPackages}{Character, a vector of system packages that will be installed
by \verb{apt-get install} before running the R worker}

\item{environment}{List, the environment variables in the container}

\item{maxWorkerNum}{Integer, the maximum worker number in a container}

\item{tag}{Character, the image tag}
}
\value{
a \code{foreachRedisContainer} object
}
\description{
Get the doRedis worker container.
}
\examples{
doRedisWorkerContainer(image = "r-base")
}
